backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asyncer==0.0.8
attrs==24.2.0
blinker==1.8.2
boto3==1.35.28
botocore==1.35.28
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gradio==4.44.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.2
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.8
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tifffile==2024.9.20
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
watchdog==5.0.2
websockets==12.0
Werkzeug==3.0.4
yarl==1.13.0
